requirements.txt
astra-toolbox==1.8
certifi==2020.12.5
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1615055117017/work
decorator @ file:///tmp/build/80754af9/decorator_1617916966915/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1594044661732/work
joblib==1.0.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282412546/work
matplotlib==3.1.1
mkl-service==2.3.0
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy==1.19.5
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
Pillow==6.2.1
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.3.2
six @ file:///tmp/build/80754af9/six_1605205335545/work
threadpoolctl==2.1.0
tifffile==2020.9.3
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
